==> Jupyter Notebooks/2_LyricsPreprocessing.ipynb <==
# Preparation

## Requirements

import pandas as pd

# Language detection
# !pip install google-cloud-translate
!pip install langdetect
from langdetect import detect, LangDetectException

import re # RegEx
import string
from sklearn.feature_extraction.text import TfidfVectorizer # To vectorize

# For stopwords
import nltk
from nltk.corpus import stopwords
# For tokenization and lemmatization
from nltk.stem import WordNetLemmatizer
from nltk.tokenize import word_tokenize

# # For punctuation
# import unicodedata

import gdown

## Dataset Import

### Original Dataframe

Index: 450127 entries, 0 to 502193

Data columns (total 43 columns):

 N°   Column                          Non-Null Count   Dtype         
---  ------                          --------------   -----         
 0   Artists                         450127 non-null  object        
 1   song                            450115 non-null  object        
 2   text                            450127 non-null  object        
 3   Length                          450127 non-null  Int64         
 4   emotion                         450127 non-null  object        
 5   Genre                           450127 non-null  object        
 6   Album                           450078 non-null  object        
 7   Release_Date                    450127 non-null  datetime64[ns]
 8   Key                             450127 non-null  object        
 9   Tempo                           450127 non-null  Int64         
 10  Loudness_db                     450127 non-null  float64       
 11  Time_signature                  450120 non-null  object        
 12  Explicit                        450127 non-null  boolean       
 13  Popularity                      450127 non-null  Int64         
 14  Energy                          450127 non-null  Int64         
 15  Danceability                    450127 non-null  Int64         
 16  Positiveness                    450127 non-null  Int64         
 17  Speechiness                     450127 non-null  Int64         
 18  Liveness                        450127 non-null  Int64         
 19  Acousticness                    450127 non-null  Int64         
 20  Instrumentalness                450127 non-null  Int64         
 21  Good_for_Party                  450127 non-null  bool          
 22  Good_for_Work_Study             450127 non-null  bool          
 23  Good_for_Relaxation_Meditation  450127 non-null  bool          
 24  Good_for_Exercise               450127 non-null  bool          
 25  Good_for_Running                450127 non-null  bool          
 26  Good_for_Yoga_Stretching        450127 non-null  bool          
 27  Good_for_Driving                450127 non-null  bool          
 28  Good_for_Social_Gatherings      450127 non-null  bool          
 29  Good_for_Morning_Routine        450127 non-null  bool          
 30  Similar_Artist_1                450127 non-null  object        
 31  Similar_Song_1                  450110 non-null  object        
 32  Similarity_Score_1              450127 non-null  float64       
 33  Similar_Artist_2                450127 non-null  object        
 34  Similar_Song_2                  450112 non-null  object        
 35  Similarity_Score_2              450127 non-null  float64       
 36  Similar_Artist_3                450127 non-null  object        
 37  Similar_Song_3                  450123 non-null  object        
 38  Similarity_Score_3              450127 non-null  float64       
 39  year                            450127 non-null  int32         
 40  month                           450127 non-null  int32         
 41  day                             450127 non-null  int32         
 42  decade                          450127 non-null  int32         

### Filtered Import

interest_columns = [
    'Artists',
    'song',
    'text',
    'emotion',
    'Genre',
    'Album',
    'Release_Date',
    'Explicit',
    'Popularity',
    'Danceability'
    ]

# # # So that Pandas gets only the columns of interest
# # songs = pd.read_csv('/content/clean_songs_df.csv', usecols=interest_columns, engine='python')

# songs = pd.read_csv('/content/clean_songs_df.csv', engine='python')

# url = 'https://drive.google.com/file/d/1w7Kqc58mNQdzVe5QgHuDYnD5CbCygZgG/view?usp=drive_link'
# path = 'https://drive.google.com/uc?export=download&id='+url.split('/')[-2]
# songs = pd.read_csv(path, engine='python')

url_file = 'https://drive.google.com/file/d/1w7Kqc58mNQdzVe5QgHuDYnD5CbCygZgG/view?usp=drive_link'
file_id = url_file.split('/')[-2]
url = f"https://drive.google.com/uc?id={file_id}"

output = "clean_songs_df.csv"
gdown.download(url, output, quiet=False)

songs = pd.read_csv("clean_songs_df.csv")

songs.info()

songs.head()

len(songs)

## Text Preprocessing

### Lyrics Cleaning and Lemmatization

### Language

# Define stopwords
nltk.download('stopwords')
stop_words = set(stopwords.words('english'))

# Define lemmatizer
lemmatizer = WordNetLemmatizer()

# Define puntuation,
# nltk.download('punkt')
nltk.download('punkt_tab')
nltk.download('wordnet')
nltk.download('stopwords')

# def clean_text(text):
#     text = re.sub(r'\[.*?\]', '', text) # Get rid of stuff like "[Verse 1]", "[Intro]" or "[Artist]"
#     text = text.lower() # Lowercase
#     text = re.sub(f"[{re.escape(string.punctuation)}]", "", text) # Get rid of punctuation
#     tokens = text.split() # Split to get rid of stopwords
#     tokens = [word for word in tokens if word not in stop_words] # Get rid of stopwords
#     return " ".join(tokens) # Rejoin to make the text again

# # nltk.download('punkt')

# def clean_and_lemmatize(text):
#     text = re.sub(r'\[.*?\]', '', text)  # Remove tags such as [Verse 1]
#     text = text.lower()
#     text = re.sub(f"[{re.escape(string.punctuation)}]", "", text)
#     tokens = word_tokenize(text)
#     tokens = [word for word in tokens if word not in stop_words]
#     tokens = [lemmatizer.lemmatize(word) for word in tokens]
#     return " ".join(tokens)

# def remove_punct(text):
#     # Get rid of any punctuation characteraccording to Unicode
#     return ''.join(ch for ch in text if not unicodedata.category(ch).startswith('P'))

# def clean_and_lemmatize(text):
#     text = re.sub(r'\[.*?\]', '', text) # Get rid of stuff like "[Verse 1]", "[Intro]" or "[Artist]"
#     text = text.lower() # Lowercase
#     text = remove_punct(text) # Get rid of punctuation
#     tokens = word_tokenize(text) # Split to tokenize and get rid of stopwords
#     tokens = [word for word in tokens if word not in stop_words] # Get rid of stopwords
#     tokens = [lemmatizer.lemmatize(word) for word in tokens] # Lemmatization
#     return " ".join(tokens) # Rejoin to make the text again

def remove_punct_tokens(tokens):
    # return [t for t in tokens if not unicodedata.category(t).startswith('P')]
    return [t for t in tokens if t.isalpha()]

def clean_and_lemmatize(text):
    text = re.sub(r'\[.*?\]', '', text) # Get rid of stuff like "[Verse 1]", "[Intro]" or "[Artist]"
    text = text.lower() # Lowercase
    tokens = word_tokenize(text) # Split to tokenize and get rid of stopwords
    tokens = remove_punct_tokens(tokens) # Get rid of punctuation
    tokens = [word for word in tokens if word not in stop_words] # Get rid of stopwords
    tokens = [lemmatizer.lemmatize(word) for word in tokens] # Lemmatization
    return " ".join(tokens) # Rejoin to make the text again

stop_words

print(clean_and_lemmatize("You're not going to [Verse 1] believe this!"))

# test = songs.iloc[:50].copy()

# # test['clean_text'] = test['text'].apply(clean_text)
# test['clean_text'] = test['text'].apply(clean_and_lemmatize)

# print(test['clean_text'])

#### Google Translate API (momentarely not used!)

# !gcloud auth application-default login

# from google.cloud import translate_v2 as translate

# def get_language(batch_texts):
#     translate_client = translate.Client()
#     results = translate_client.detect_language(batch_texts)
#     # results is a list of dictionaries, one per text
#     return [result['language'] for result in results]

#### Langdetect

# def get_language_local(batch_texts):
#     return [detect(text) for text in batch_texts]

# batch_size = 5
# sample_batch = songs['text'].iloc[:batch_size].str[:100].tolist()
# idiomas = get_language_local(sample_batch)
# print(idiomas)

# test = songs.iloc[:5].copy()

# # for song in test:
# #     test['language'] = detect(song['text'])

# test['language'] = test['text'].apply(lambda x: detect(x[:100]) if isinstance(x, str) and len(x) > 0 else 'unknown')
# # print(test[['text', 'language']])

# test

# songs['language'] = songs['text'].apply(lambda x: detect(x[:100]) if isinstance(x, str) and len(x) > 0 else 'unknown')

def get_language(text):
    i = 0

    try:
        if isinstance(text, str) and len(text.strip()) > 0:
            i += 1
            print(f'{i} Detected!')
            return detect(text[:100])
        else:
            return 'unknown'
    except LangDetectException:
        return 'unknown'

### Text Cleaning

# Define languages
songs['language'] = songs['text'].apply(get_language)

# Get rid of non-english lyrics
songs = songs.loc[songs['language'] == 'en']

songs.info()

# Clean the lyrics
songs['clean_text'] = songs['text'].apply(clean_and_lemmatize)

songs.info()

# songs.sample(10)

# Get rid of songs with an empty "song" field
songs = songs.loc[~songs['song'].isnull()]

songs.info()

## Set index

songs['song_id'] = songs.index
songs.set_index('song_id', inplace=True)

## Save Dataframe

# songs.to_csv('songs_clean_text.csv', index=False)
songs.to_csv('songs_clean_text.csv', index=True)

from google.colab import files
files.download('songs_clean_text.csv')
