==> Jupyter Notebooks/4_UniteEmbeddingBatches.ipynb <==
import pandas as pd

# To import data
import gdown

# URLs
urls = [
    'https://drive.google.com/file/d/1R6NaUFr6aSaZDKR3CjPevKB9WuF0c_p-/view?usp=sharing',
    'https://drive.google.com/file/d/1K2ip85ygPu4FOev03ppakF-_iyAl4K9E/view?usp=sharing',
    'https://drive.google.com/file/d/1zWaREPqEE58EW2RQNhlAeEcCUaUtJhJ0/view?usp=sharing',
    'https://drive.google.com/file/d/1VI1lUYYXqr5nnV5zxgxLL1_UszAuL7_e/view?usp=sharing',
    'https://drive.google.com/file/d/1R5nUyozgCUORmwkdCxAmzQJIwAEDCPnd/view?usp=sharing',
    'https://drive.google.com/file/d/1HNGkkcmTOogH69vIfmJAtAyJm4RM3fgc/view?usp=sharing',
    'https://drive.google.com/file/d/1eAwn1MUqIFYK7uOs5t4eMmXKgdOjpkPZ/view?usp=sharing',
    'https://drive.google.com/file/d/1P0xBqcTignQ5L70aQgVbaBEenZ-yJ0D7/view?usp=sharing',
    'https://drive.google.com/file/d/1e2jVDLI5Ulsb-czLC1BW5iuCZKP2PvoZ/view?usp=sharing'
]

# Import divisions dataset
i = 0
for url in urls:
    i += 1
    division = url.split('/')[-2]
    url_division = f"https://drive.google.com/uc?id={division}"

    output_division = f'division_{i}.csv'
    gdown.download(url_division, output_division, quiet=False)

len(urls)

# Lista de archivos
files = [f'division_{i}.csv' for i in range(1, len(urls) + 1)]
# Leer y concatenar
complete_df = pd.concat([pd.read_csv(file) for file in files], ignore_index=True)

# complete_df.info()

# complete_df['Q1'].info()

# complete_df.dtypes

# Save results
complete_df.to_csv("complete_embedded_df.csv", index=False)

# Descargar el archivo
from google.colab import files
files.download("complete_embedded_df.csv")
