==> Jupyter Notebooks/5_Cosine_Similarity_sample.ipynb <==
test = True

# Step 1: Import libraries
import pandas as pd
from sklearn.metrics.pairwise import cosine_similarity

# Step 2: Load my embeddings that are already done
if test == True:
   	url_songs = 'https://drive.google.com/file/d/1OYu-bt9GQQ_q1SJ1295w4RmzH-nXAcs7/view?usp=sharing'
else:
    url_songs = 'url de todo junto'
path_songs = 'https://drive.google.com/uc?export=download&id='+url_songs.split('/')[-2]
song_embeddings = pd.read_csv(path_songs)

url_events = 'https://drive.google.com/file/d/1qYjdla35SWFpPFBi24uIJXzErMIGWQUi/view?usp=sharing'
path_events = 'https://drive.google.com/uc?export=download&id='+url_events.split('/')[-2]
event_embeddings = pd.read_csv(path_events)

event_embeddings.head(1)

song_embeddings.head(1)

# Step 3: Make sure that we're only using the vectors (384 numeric columns)
# If there are columns like 'song' or 'event', we eliminate them
song_vectors = song_embeddings.drop(columns=['song_id']).values
event_vectors = event_embeddings.drop(columns=['EventCode']).values

song_vectors

# Step 4: Calculate the cosine similarity matrix
similarity_matrix = cosine_similarity(song_vectors, event_vectors)

# Step 5: Create df
similarity_df = pd.DataFrame(similarity_matrix,
                             index=song_embeddings['song_id'],
                             columns=event_embeddings['EventCode'])

similarity_df.shape[0], similarity_df.shape[1]

similarity_df.head(10)

similarity_df.describe()

similarity_df.head(1)

# similarity_df.set_index('EventCode', inplace=True)

# # So that the index gets saved as well
# similarity_df.set_index('song', inplace=True)
# Step 6: Save results
# similarity_df.to_csv("similarity_df.csv", index=False)
similarity_df.to_csv("similarity_df.csv", index=True)

# Descargar el archivo
from google.colab import files
files.download("similarity_df.csv")

similarity_df.head(5)

import pandas as pd
import numpy as np

# Aplanamos la matriz manteniendo los índices
stacked = similarity_df.stack()

# Obtenemos los 4 valores más bajos
lowest = stacked.nsmallest(4)
# Obtenemos los 4 valores más altos
highest = stacked.nlargest(4)

# Mostramos resultados
print("🔻 Canciones con menor similaridad:")
print(lowest)

print("\n🔺 Canciones con mayor similaridad:")
print(highest)

# Asegurarse de que el índice sea simple
similarity_df_reset = similarity_df.reset_index()

# Obtener song_index correcto antes de apilar
similarity_df_reset['song_index'] = similarity_df_reset.index

# Apilar matriz para tener song, EventCode, similarity
stacked_df = similarity_df_reset.set_index(['song_index', 'song_id']).stack().reset_index()
stacked_df.columns = ['song_index', 'song_id', 'EventCode', 'similarity']

# Filtrar canciones vacías o nulas
stacked_df = stacked_df[stacked_df['song_id'].notna() & (stacked_df['song_id'] != '')]

# Top 4 más bajos
lowest = stacked_df.nsmallest(15, 'similarity')
# Top 4 más altos
highest = stacked_df.nlargest(15, 'similarity')

# Mostrar ambos
print("🔻 Canciones con menor similaridad:")
print(lowest[['song_index', 'song_id', 'EventCode', 'similarity']])

print("\n🔺 Canciones con mayor similaridad:")
print(highest[['song_index', 'song_id', 'EventCode', 'similarity']])

url_originaldf = 'https://drive.google.com/file/d/1nIcwGfU9WB2UCcflYzEhr6c9VU_-F6WH/view?usp=sharing'
path_originaldf = 'https://drive.google.com/uc?export=download&id='+url_originaldf.split('/')[-2]
#original_df = pd.read_csv(path_originaldf)

import gdown
output_original = 'original.csv'
gdown.download(url_originaldf, output_original, quiet=False)

# To import data
import gdown
# Import songs dataset
url_songs_file = 'https://drive.google.com/file/d/1pp6AwVMt2ROIMX96YPqkdmyDmMxA2G6w/view?usp=sharing'
songs_id = url_songs_file.split('/')[-2]
url_songs = f"https://drive.google.com/uc?id={songs_id}"
output_songs = "songs_clean_text.csv"
gdown.download(url_songs, output_songs, quiet=False)
# songs_lyrics = pd.read('/content/songs_clean_text.csv', encoder='python')
songs_lyrics = pd.read_csv("songs_clean_text.csv")

songs_lyrics.head(1)

# Test

aux_df.loc[aux_df['song_id'].isin([329144,375321,168540,66986,110350,74875])]

aux_df = songs_lyrics[['song_id','Artists','song','text']]

aux_df.loc[aux_df['song_id'].isin([28924, 329144, 375321, 168540, 66986, 110350])]

aux_df['Artists'].unique()

pd.DataFrame({'Artist': aux_df['Artists'].dropna().unique()})

pd.DataFrame({'Artist': aux_df['Artists'].dropna().unique()}).to_csv("unique_artists.csv", index=False)

from google.colab import files
files.download("unique_artists.csv")

aux_df.loc[aux_df['Artists'].str.contains('Bob Marley', case=False, na=False)]

aux_df.loc[aux_df['Artists'] == 'Sex Pistols']

aux_df.loc[aux_df['song'].str.lower() == 'b.y.o.b.'.lower()]

aux_df.loc[aux_df['Artists'].str.contains('System of a down', case=False, na=False)]

aux_df.loc[aux_df['song'].str.lower() == 'god save the queen'.lower()]

list_99249 = [-0.034050293,0.03137639,-0.013751942,-0.04142069,-0.016881144,0.14396207,0.06138286,-0.027261928,0.010815433,0.00804828,0.022114355,-0.043791782,0.03231131,-0.042822786,-0.025716156,0.04449826,-0.04428636,0.041166153,-0.013514375,-0.027613996,0.01138959,0.09246151,-0.06569752,0.061645348,-0.09192335,0.09060175,-0.084854186,-0.01173295,0.017059814,-0.063203774,0.06530878,0.0026319847,0.043044087,-0.023789683,-0.10153499,0.041341748,0.03860937,0.05641478,0.008962404,0.017046839,-0.008276726,-0.009965662,-0.003932244,-0.10426438,0.041773852,-0.0032297354,-0.046868388,-0.08679829,0.09105562,-0.068782516,-0.03623025,-0.030632475,0.037454814,0.027204186,0.03318816,0.00028928602,-0.006159352,0.05143739,0.06539457,-0.02314305,-0.08402618,-0.04172047,0.040950373,0.016712403,-0.0008403481,0.04154137,0.014583849,0.03588764,-0.02099063,0.10772906,-0.10608459,0.019270623,0.045837108,-0.050951425,-0.045018625,0.029819107,-0.037214953,-0.11653151,-0.023512887,-0.0103433095,-0.00024740683,-0.03042464,-0.041948024,-0.010703422,-0.01843151,-0.031139463,0.025207281,0.06479881,0.0059361164,-0.13309018,-0.12331023,0.05179912,0.043559063,0.052333783,-0.0655715,0.0003837254,-0.0713499,-0.05468476,-0.10081084,0.038446616,0.07550407,0.015458884,0.040233698,-0.029906012,0.076924875,-0.039251633,-2.412478e-05,-0.00014602185,0.014747939,0.044257462,0.052861534,-0.026083073,-0.043383032,-0.017370228,0.06137633,0.039795246,0.009236983,-0.027659865,-0.06946286,0.002949692,0.03115475,-0.018353295,0.037884764,0.024932936,0.068318434,-0.019304793,-0.0424122,1.9865302e-32,0.018705793,-0.036824297,0.02583058,0.08262696,0.10549255,-0.039497823,-0.05075181,0.039553043,-0.0790605,0.039515596,0.0373921,0.017195247,0.02232457,-0.016228838,-0.0709267,-0.032229908,-0.03557559,-0.06709576,-0.055612583,-0.016521445,-0.040962096,0.0141552,-0.031512693,0.009232702,0.0102971075,-0.012591756,0.01919032,0.018927384,0.0382714,-0.019578923,0.02757438,0.015560212,-0.022566374,-0.031094376,0.09647798,-0.0032678358,0.016398523,0.03149763,0.0017265086,-0.01884797,-0.09179129,0.024978595,-0.06142727,-0.049721364,0.045252487,0.09229746,-0.00703375,-0.029945627,-0.0026376809,0.018016243,-0.049018443,0.05768944,0.020539986,0.03505255,0.070188135,-0.027797801,-0.04370753,0.06842393,0.026489975,-0.016558824,-0.116380095,0.054081995,0.032643586,0.053052768,-0.058430668,-0.037609514,-0.045757923,-0.05989579,-0.04144516,0.018348183,0.03465625,-0.047962766,-0.02734107,0.029301697,0.009709546,-0.081263095,0.0033173005,-0.041995235,-0.061402082,0.028311443,-0.002905657,-0.050486762,-0.043103643,-0.055218887,0.049267408,-0.07951621,0.0822133,-0.11264287,-0.08479832,-0.09332387,-0.037600327,0.06130277,0.10054148,-0.118574865,-0.06399025,-1.852707e-32,0.033021603,0.05800855,0.044972483,0.09978136,0.066900216,0.03845475,0.002663889,0.009857077,0.06984195,0.10348434,0.052501287,0.02410436,-0.06118183,-0.050000414,0.103911914,-0.12753193,0.032101877,0.04126991,0.03834541,0.063877665,-0.057708524,-0.025400823,-0.16042486,0.0035786256,-0.005844925,0.07131929,0.0040733363,0.024950868,0.013744321,-0.01118205,-0.03988953,-0.039146315,-0.11384013,0.05186336,-0.06679625,-0.027651308,-0.0642197,-0.016506892,-0.07794153,-0.0057960483,-0.094985716,0.012911317,-0.07050482,-0.008875881,-0.04993044,-0.03415073,-0.0096979765,0.06566856,-0.06402864,-0.0124243805,-0.04878139,0.006023203,-0.018048495,-0.0117291175,-0.034716334,-0.033143878,-0.05221945,-0.009838577,-0.071317926,0.029828623,0.07708598,-0.036878858,-0.049300935,0.015156174,0.037982367,0.010336254,-0.021005133,0.08612275,0.0015484275,0.035792716,-0.013591703,-0.042126738,-0.07929611,0.012566988,0.011035394,0.07951243,-0.006715753,-0.0077079036,-0.03184463,-0.047956772,0.030891761,-0.056015752,-0.062191326,0.039528754,-0.02373612,0.04170284,-0.014027928,-0.020322366,-0.009834269,0.06530277,0.06212453,0.06883047,0.04828704,0.013794071,0.022360973,-5.0772293e-08,-0.06261427,0.007901777,-0.00789952,0.022598311,0.027381148,0.021819143,0.028834097,-0.08654715,-0.015189695,0.029260218,0.016361507,0.02653181,0.029089589,0.020480286,-0.025156632,0.034948688,-0.098874636,-0.05903384,-0.052960675,0.01638284,-0.036442865,-0.02281836,0.025583487,-0.06280404,-0.009602437,0.034493733,-0.0055352543,0.021488577,-0.08507873,0.0380327,0.018023612,0.021583945,-0.08693203,0.012025942,0.06489017,-0.02980435,-0.030720884,0.053801082,0.015151923,0.005863094,-0.03879846,0.080568045,0.0792231,0.036471795,-0.117712066,0.0099174995,0.0069467016,0.04615565,-0.02239937,-0.09792959,0.016370118,0.030273631,0.008490723,-0.0063098385,0.016286891,0.08899344,0.040916327,0.11609809,-0.037473872,-0.0036808143,0.013563636,-0.08411411,-0.0052897194,-0.009299591]
list_7537 = [0.0021153947,0.03318989,-0.008942032,-0.10200521,0.018403025,0.14050125,0.067003265,-0.077204555,-0.03381728,0.057866465,-0.011053394,-0.034575697,0.070296004,-0.12141645,-0.04431275,0.02223388,-0.068989865,0.04506298,-0.09611535,0.021768807,0.07930464,-0.010468588,-0.07249509,0.05203322,-0.13309474,0.020785192,-0.037724722,-0.023143701,0.04210555,-0.09888642,0.07311678,-0.031550817,0.043726124,0.056778934,-0.007428299,0.050017606,-0.022515006,0.056603067,0.01260801,-0.012433431,0.012573659,-0.06430511,-0.08081168,-0.011498415,-0.026578566,0.02892119,-0.03153637,0.005452644,0.027216243,-0.038532604,-0.043687627,-0.04007118,0.028365415,-0.0104823755,-0.019943021,-0.018134814,-0.0014916689,0.017475588,0.052776046,-0.05006528,0.0028498077,0.06385612,0.022906648,-0.020928385,-0.06547248,-0.027993532,0.027610866,0.098725855,-0.07693947,0.072340995,-0.03880921,0.020022523,0.02512313,0.012605119,-0.05695945,0.014297733,0.009301344,-0.0836163,-0.021499962,0.046352707,-0.043770816,-0.058005378,-0.021334779,0.031065447,0.0025552192,0.0075906008,-0.047075175,-0.06111775,0.05986996,-0.096526295,-0.0476602,-0.08654784,0.16198626,0.06946536,-0.06811642,0.04118104,-0.016559629,-0.11004727,-0.04339337,0.088987365,-0.027672024,0.028806195,-0.052555576,0.025418514,0.04103948,-0.04968711,-0.0023144998,0.01615108,0.007941208,0.008115236,0.036220677,0.07006925,-0.008865198,0.013627138,0.054842852,-0.0013483377,-0.13162534,-0.009165172,-0.06014335,-0.014141561,0.07570256,0.054641284,-0.04781174,0.030089885,0.0139649585,-0.013151402,0.02547176,1.4484656e-32,-0.03679723,-0.019713355,0.020053828,0.04591306,0.027468447,-0.016719345,-0.0019122374,0.05014048,-0.0115555795,-0.012345681,-0.05522711,-0.039801296,-0.046350744,-0.009915366,-0.053062264,0.03702828,-0.05096573,-0.03250036,0.029852314,0.016282838,-0.0320924,0.05971982,0.0901106,0.0013046367,0.0038026697,-0.07204391,0.036191795,-0.018096976,0.12135984,0.05867558,0.0055280356,-0.014443394,0.042936202,-0.02178373,-0.026610488,-0.026243718,-0.04168898,-0.08539473,-0.038132075,-0.0050443104,-0.058384735,0.0115324985,0.0794798,0.029003602,-0.004845127,-0.0008080374,-0.01811577,-0.023420176,-0.023350911,0.06513871,0.015926383,-0.042751513,-0.010077814,-0.010015865,-0.009332469,-0.02595544,-0.051554214,0.05704868,0.029270262,0.0337528,-0.039137896,-0.08300126,0.08381703,0.00030326945,-0.023099508,-0.08033705,-0.013762649,-0.038943067,-0.023747843,-0.011495671,-0.047900278,0.022232361,-0.0070534167,0.051622298,0.0048664315,-0.018168494,0.028843625,-0.05656905,-0.0055225375,-0.03545621,-0.011802621,-0.027538227,-0.021045662,-0.016741415,0.1678568,-0.10643921,0.001848367,-0.16215995,0.03400449,-0.009413158,-0.007127185,0.001459805,0.08555174,-0.031034917,-0.095166184,-1.334579e-32,0.033217683,-0.045052797,0.0053896876,0.1665245,-0.02101565,0.062150035,-0.07520808,0.036614057,0.009954368,0.018159062,-0.032590274,-0.005100147,0.008676002,0.009236179,0.0619381,-0.11222974,0.051894758,-0.033129204,-0.030829344,-0.017736817,-0.014691475,-0.03834442,-0.068258636,0.024063464,-0.019763188,0.05826612,-0.07882496,0.058724504,0.031244127,0.0380239,0.0033046326,-0.012204751,-0.21122871,0.023130603,-0.023302581,-0.028397655,0.013918126,0.012009352,0.018894052,0.0060336892,-0.08088742,-0.071106076,0.05804493,-0.021342633,0.0018475344,-0.038560078,0.015212925,0.07779844,0.13163035,0.030251943,-0.02062222,-0.032038156,0.007124064,-0.004234955,0.01520181,-0.02264859,-0.018372389,-0.039251506,-0.033629563,0.019655516,0.010611105,-0.02899715,0.0033743356,0.024067447,0.00082875724,0.05897555,0.025077777,0.088274255,0.06050069,0.027525378,-0.023545133,-0.011126287,-0.19049525,-0.032279097,-0.037433002,0.003174241,-0.047638107,-0.07102076,-0.029415447,-0.0081641385,0.038636073,0.019133305,-0.05605936,-0.028319534,0.0050079008,-0.012629553,0.07698616,0.017815739,0.06454171,0.017174736,0.046371695,0.06977449,0.07389556,-0.08444051,0.03592445,-4.4315573e-08,-0.0071470696,0.058892675,-0.010438668,0.008634069,-0.00028305769,-0.05487132,-0.012772828,-0.082726076,-0.01183241,-0.019332299,0.045406077,0.056721315,0.014045101,-0.031993467,-0.0068882843,-0.0044353893,-0.05866896,-0.016223334,-0.036915906,-0.063956805,0.041631356,-0.016248839,0.045973256,-0.060056604,0.006097757,-0.053304143,0.03043464,0.04127922,0.007130921,0.07322654,0.03276152,0.036005203,-0.063603595,-0.0091015585,0.050865617,-0.008987196,0.0012523871,0.031248767,0.08628371,0.01678154,-0.0026646627,0.1182825,0.021835666,0.0026749235,-0.051491905,-0.012817219,0.039571162,0.032678697,0.06735471,-0.017346635,-0.0097495755,0.039354406,0.061061576,0.059025988,0.02172503,0.023225952,-0.06052226,0.041489664,0.066886835,-0.004893346,0.038372606,-0.049092084,0.013594287,-0.025754116]
list_6028 = [-0.020540081,0.019680332,0.06903121,0.02407581,0.12205851,0.03316073,0.062633306,-0.07761171,0.0061104787,0.01731928,-0.028459102,-0.11203416,-0.004718419,-0.04700947,0.018508248,0.056422725,-0.008105146,-0.047164142,-0.14438248,-0.00025508346,-0.012011631,0.08140727,-0.089781195,0.036320925,-0.10726231,0.09859118,-0.107046075,0.017768478,-0.026248455,-0.081459045,0.028851617,0.010207619,-0.026157966,0.020724816,0.063952796,0.023560306,0.0024668635,-0.041675963,-0.053676926,0.0055701286,0.024185244,-0.064216375,0.005101866,0.015377124,-0.051921126,0.016932178,-0.030292228,-0.027321944,0.03097568,-0.059565447,-0.0014141374,-0.12688887,-0.061795328,0.00023293651,-0.022573914,-0.017411292,0.025331158,-0.05340946,0.05860428,-0.03927626,0.015628215,-0.018459642,-0.051307887,0.028746229,0.06187629,-0.023652755,-0.020481646,0.095302284,-0.011341148,0.026578058,-0.05515542,0.0888291,-0.03732633,0.06745504,-0.05286172,0.056923375,0.026892692,-0.07058041,-0.0025471663,0.03439703,0.0021049187,-0.029269079,-0.06845856,0.03290128,-0.050636508,0.044964157,-0.045047082,0.025779175,0.010870673,-0.077426106,-0.0622058,-0.07329289,0.12482464,0.09114239,-0.10453572,0.0074808323,0.0030360497,-0.105361715,-0.07309574,0.07968823,0.05229558,0.042620353,0.022225915,-0.006376068,-0.006270501,-0.036473732,-0.05870913,0.008013643,-0.04461606,-0.039050255,-0.05743672,-0.016256733,-0.03987486,-0.06675341,0.06010529,-0.046899255,0.015511699,-0.0091360435,-0.06081896,-0.009627145,-0.021088878,-0.0018946644,0.017878043,0.030219002,0.03795281,-0.009538108,0.024014093,1.4561377e-32,0.09441423,-0.015076298,0.057324342,0.051999837,0.16650872,-0.015420996,-0.059606895,0.001820109,-0.04950634,0.03286181,-0.031968914,-0.018201586,-0.08754061,0.0004210718,-0.007922999,-0.04521079,-0.011138699,-0.023696825,-0.018459646,0.05959014,-0.056141492,0.04116087,-0.0035918932,-0.06429529,-0.016935166,-0.09965224,0.04220467,-0.024075573,0.02724876,0.010439843,0.06263804,-0.0041615735,0.01912536,-0.0060647563,-0.017169023,0.010887435,-0.04001253,0.005255792,-0.06621884,-0.062103625,-0.04624745,0.024037814,-0.011288612,0.022408761,0.10164497,-0.0013405546,0.032918293,0.042501904,0.020806672,0.016973337,0.0076171327,-0.0032717003,-0.023630695,-0.056267753,0.03246791,-0.003505931,0.05966099,0.01937268,3.6125268e-05,-0.013906652,0.027801277,-0.01818533,0.0042603742,-0.04587849,0.03541225,-0.057801478,0.061258156,-0.051716004,-0.016343268,0.004629253,-0.004031645,-0.04387036,0.03196384,0.053087804,0.07726784,0.00056008215,0.051243093,-0.08217829,-0.025962485,0.047698326,-0.01859506,-0.1131161,-0.009804106,-0.025358293,0.034171227,-0.03070983,0.039042,-0.04405204,-0.009078484,-0.04506283,-0.02366294,-0.05514051,0.086709395,-0.08126891,-0.06086621,-1.3624633e-32,0.042242046,0.026844352,-0.002538158,0.10506873,0.09338047,-0.044710446,-0.051790588,0.032928675,0.11933538,0.13839443,-0.0067860023,0.048549213,0.022902047,-0.023837097,0.08505486,-0.057003297,0.1119283,0.003981923,-0.0074088303,0.02975624,-0.066715434,0.017911978,-0.10244726,0.09694754,-0.04614087,0.07286901,0.083627775,0.02208351,-0.045664728,-0.028562808,-0.06661327,0.03174329,-0.06252223,0.05361623,-0.05442444,-0.047888078,0.030894393,-0.053178657,-0.07988209,-0.051990665,-0.043594822,-0.0042049945,0.011493784,0.029542161,-0.08340196,-0.025143962,-0.0774775,0.039688986,-0.004285232,0.0670293,0.016496733,-0.018175825,-0.05162914,0.07664017,0.0365605,-0.033460803,-0.0303975,-0.03910415,-0.02652619,0.05547059,0.021435924,0.009652342,-0.059079524,-0.018049123,-0.009867085,-0.08291981,-0.048509266,0.05856956,-0.033069246,0.010986446,0.012989642,-0.033977482,-0.14210057,-0.07645521,-0.008626025,-0.017494563,0.022660049,0.0038945244,-0.0025390484,0.05061354,-0.009213294,0.04166449,0.0045978017,0.042090915,0.09765431,-0.04882328,0.066645205,0.0219419,0.03580363,0.053067125,0.032167993,0.063265696,0.020975124,0.06795096,-0.03664474,-4.1791083e-08,0.028897502,0.046260003,0.0002898013,-0.0070231888,0.008765665,-0.09125698,0.12523584,0.05212034,-0.009319894,0.045614943,0.053774964,0.051206358,-0.0016014552,0.053386357,0.03571552,0.0046058726,-0.05004683,-0.041386392,-0.078795,-0.043030746,0.034995306,0.0076411488,0.04975457,-0.0186695,0.022706842,0.013233125,0.0009446598,0.027735556,0.092118196,0.08757676,0.04738384,0.029429033,-0.05877267,-0.032201454,-0.104741395,0.0016777667,0.008749096,0.025829922,-0.07561775,-0.0060316217,-0.010497222,0.031410925,0.03339062,-0.0026724203,-0.0638565,-0.057580218,-0.04188641,-0.0058979155,-0.025200384,-0.035011332,-0.035995867,-0.025630737,0.05299698,0.037186123,0.025252473,-0.027890608,-0.022838673,0.06105677,0.018887738,0.011643123,0.063767314,-0.10958052,0.0071754106,0.035260547]
list_141539 = [0.016682442,0.03720659,-0.010714134,-0.053325616,0.03921881,0.0745829,0.099119,-0.038038872,0.030028112,0.009810958,-0.0030858628,-0.08811174,0.04948246,-0.048324224,0.004136675,0.051301565,-0.09019148,-0.011251333,-0.07551937,0.00391799,0.05870692,0.017182257,-0.029577786,-0.011610938,-0.14689684,0.03530336,-0.0010604616,-0.031795066,0.033773314,-0.030695803,0.061602205,-0.0044585397,-0.011450523,0.017894119,0.012750138,0.009642422,0.0011683294,0.07483389,0.06230273,-0.10114648,-0.029266758,-0.08204244,-0.049935788,-0.05652533,-0.010012683,0.0391636,-0.04914081,-0.005901621,0.042349268,-0.08632252,-0.06339926,-0.11148357,0.0025155277,-0.07366392,-0.01152079,-0.03331019,-0.0054875677,0.008358454,0.024362195,-0.073408976,-0.024960252,-0.011563801,0.057032265,-0.012654211,0.025300493,0.025824491,-0.025202174,0.09475584,-0.067818075,0.04300054,-0.051479366,-0.008321829,-0.0098443795,-0.030689968,-0.052200615,-0.013613074,0.028818434,-0.038038433,-0.031881087,0.0200705,-0.043786258,-0.03483167,0.0074758576,0.03808863,-0.027979976,0.036266528,-0.03674281,-0.040513918,0.029450966,-0.020464377,-0.02031732,-0.12514977,0.13601503,0.08985586,-0.01859291,0.07786474,-0.015551691,-0.07308232,-0.040642936,0.091251835,0.035939854,0.03499184,-0.022127712,0.040204808,0.030665971,-0.06903544,0.006085158,-0.0038521695,0.0005911846,-0.018951096,0.0005032641,0.042931806,-0.021337211,-0.0697595,0.046009984,-0.042305913,-0.07291131,0.017549906,-0.026732298,-0.008236596,-0.022548782,0.067012876,-0.04617031,0.03383435,0.0033420178,-0.07340615,-0.046103537,1.9800462e-32,0.022238526,-0.035147537,0.029299134,0.026215795,0.03616986,0.06781955,-0.043308716,0.048483055,0.022495165,-0.06756886,-0.07353487,-0.05491904,-0.108257554,0.08544594,-0.06080598,0.002981662,0.011718236,-0.07645596,0.006386343,0.016886562,0.026617436,0.076952934,0.04406567,-0.043700416,0.037602726,-0.013947094,-0.009161576,-0.045441475,0.07367347,0.05240706,-0.0585889,0.03532909,-0.029152744,-0.05163664,0.021138487,-0.017974647,-0.016585158,-0.09469009,-0.06358422,-0.08772427,-0.017232938,0.081587106,-0.01645579,0.022894584,0.014534336,0.0015267803,0.051089626,0.024732662,-0.03413953,0.07923722,-0.010547642,-0.0009637261,-0.06886397,-0.09436834,-0.024500206,-0.058227558,-0.040489577,0.061414324,0.029002685,-0.0006399695,-0.024494141,-0.05589973,0.051933147,0.027086675,0.01513826,-0.031167673,0.047929976,-0.04203654,0.0062797056,0.08438623,-0.0456867,0.01582047,0.018327525,0.07913505,0.012471358,0.011626638,-0.027937869,-0.046503983,-0.041389883,0.010826584,0.0073750927,-0.016391734,-0.03785337,-0.029544571,0.17029777,0.020562975,0.05163125,-0.13356274,0.020562105,-0.019099653,-0.031384915,0.008116111,0.1103653,-0.016415883,-0.18443626,-1.8619437e-32,-0.0019738693,-0.01744608,-0.0045843557,0.07524489,0.037015375,-0.024186172,-0.06927586,-0.0055961837,0.10735862,0.07753757,-0.025410598,-0.017639179,0.07327657,0.009669975,0.036350198,-0.101536736,0.064807706,-0.081431836,-0.030796142,0.026246127,-0.06079737,0.016092526,-0.09985704,0.037861608,-0.052485734,0.07091757,-0.03368464,0.04907355,-0.0024837141,-0.0017261979,-0.0364283,0.030161433,-0.12373229,0.060511764,-0.0055114124,-0.065078676,0.019429615,-0.01729895,0.029615147,0.042872153,-0.03565794,-0.0454043,-0.051448833,-0.013674482,-0.042366832,-0.022131937,-0.06257835,0.050285615,0.13367058,0.0058494555,0.034090508,-0.008570083,-0.08055084,-0.042131647,0.022883825,-0.057017528,-0.0052010184,-0.07208776,-0.030777894,0.05346897,-0.028258586,0.003943305,0.0220678,0.05009763,-0.020119498,0.06032707,-0.029922264,0.055479027,0.043053295,-0.020654565,0.06450965,-0.024503263,-0.1548589,-0.028653791,-0.042345352,0.0033602323,-0.0034656571,-0.05136758,0.03752502,0.008451809,0.017922796,0.0017502009,-0.031850632,0.016188895,0.0013502089,-0.030114658,0.0006997551,0.01367451,0.040345352,0.01613868,0.05213201,0.030833155,0.029974094,0.03710387,-0.01121775,-4.9954952e-08,-0.03884893,0.04852735,0.021823762,0.081658885,0.045904804,-0.058677465,-0.034120847,-0.09565531,-0.00926128,-0.031003602,0.0811415,0.0654879,-0.04134794,0.009618939,0.022941574,0.040146377,-0.07209558,-0.049675155,-0.01926265,-0.057683785,0.053763382,-0.011640235,0.024338132,-0.045478303,0.038199022,-0.020244364,0.003762528,0.044594638,-0.005685202,0.09589237,0.007732656,-0.027126595,-0.06692572,-0.024898065,0.022073604,0.00012047581,-0.014160623,-0.022434043,0.054557703,0.005328351,0.06567878,0.08895481,0.026329253,0.034371074,-0.018102651,-0.08673749,0.008785962,-0.017798468,0.028414596,-0.08010909,-0.008509874,0.079718165,0.10113734,0.07899253,0.049279798,-0.036448184,-0.017929668,0.03897835,0.0007788474,-0.034763895,0.10807725,-0.07526822,0.06749407,-0.05715499]
list_259669 = [0.017814444,0.040238414,-0.010711211,-0.056180038,0.04618261,0.09695441,0.122836545,-0.043033168,0.034372162,0.026560655,0.008969705,-0.0894987,0.04551158,-0.056116734,0.0067998115,0.053979736,-0.086530104,-0.029319283,-0.11347847,0.02069619,0.056012463,0.0047867247,-0.02475965,-0.017344872,-0.12175997,0.048407722,-0.017369341,-0.045312505,0.02193383,-0.051666755,0.07253443,0.011558984,-0.02484354,0.0430284,0.03032073,-0.00602298,0.0193709,0.07659149,0.07611757,-0.10027532,-0.046997927,-0.075648285,-0.054957245,-0.06100853,-0.023149747,0.035801504,-0.046878148,-0.00756272,0.03903706,-0.07350358,-0.058984164,-0.10545612,0.007439161,-0.05564102,0.0017213508,-0.04271825,0.0063539953,0.009008048,0.012019775,-0.07849133,-0.040371567,-0.009938049,0.041011572,-0.02290098,0.0264354,0.040943082,-0.025485849,0.07871818,-0.06782246,0.013024591,-0.048802614,-0.0029696499,-0.014345884,-0.04205167,-0.065455854,-0.007703432,0.047911074,-0.055977788,-0.04233743,0.022234058,-0.046243142,-0.048917886,0.014053164,0.046796698,-0.018777192,0.052584156,-0.08154142,-0.033606812,0.008129532,-0.014904687,-0.022958217,-0.11664087,0.12264952,0.083694234,-0.0031445918,0.07619251,-0.042601667,-0.06923506,-0.024820412,0.104922004,0.016410751,0.04204943,-0.005711868,0.020431211,0.029670019,-0.075128935,0.019162139,-0.0069389152,-0.017225115,-0.018891871,0.0034050555,0.040658526,-0.013887664,-0.080521405,0.021156125,-0.059864726,-0.078993745,-0.0034259087,-0.0662664,0.0025812811,-0.0051804683,0.06485188,-0.052320167,0.03350335,-0.0064928904,-0.07773686,-0.018695435,2.0698334e-32,0.0161902,-0.026102403,0.028506592,0.032740198,0.033114485,0.08296902,-0.041605163,0.040865242,0.034918327,-0.07792701,-0.08586114,-0.07321023,-0.11412879,0.08336638,-0.055209145,-0.0023418805,0.015432699,-0.073658474,0.0069943573,0.00974848,0.01823106,0.076393425,0.03575709,-0.06064548,0.026394136,-0.016906662,-0.000758456,-0.02088679,0.064242035,0.054712407,-0.049088992,0.052721348,-0.019563854,-0.052584995,0.023498217,-0.0076382263,-0.028730944,-0.08584192,-0.055226754,-0.0856964,0.0048629045,0.078042254,-0.010592535,0.06870001,-0.005690094,0.008214924,0.05742595,0.013705343,-0.021983854,0.054872207,0.00057995494,0.013535128,-0.07058551,-0.08575443,-0.00036974132,-0.05361557,-0.013339051,0.041203987,0.023805972,0.038524352,-0.002446725,-0.047197122,0.038262,0.022216825,0.029106239,-0.04230201,0.056802124,-0.04394036,-0.016454998,0.07776023,-0.032796018,-0.0024749583,0.035957824,0.11130875,-0.0014926433,0.010044934,-0.00021280623,-0.041038968,-0.022257559,-0.01875584,0.025379436,-0.029277883,-0.034466006,-0.019729199,0.14380792,0.0056847003,0.034238063,-0.1294241,0.0021905862,-0.03337379,-0.025196312,0.0035137532,0.12263948,-0.028783504,-0.18890953,-1.9703907e-32,-0.017490307,0.0034123834,-0.018159552,0.055326484,0.024987666,0.022954963,-0.079878435,-0.017143885,0.081359535,0.0878722,-0.017939905,-0.013066768,0.066367544,0.0037118432,0.036491815,-0.097766064,0.076914944,-0.08197539,-0.02348229,0.015790414,-0.053841658,0.03476524,-0.08723168,0.052845806,-0.07040435,0.074139394,-0.0412733,0.047688488,0.011690445,0.009927878,-0.012870276,0.024410464,-0.13386098,0.05535697,-0.0018843776,-0.06121859,0.03296365,-0.015009881,0.019983266,0.014121399,-0.05518422,-0.056516312,-0.0076337275,-0.0010120892,-0.045566794,-0.040035974,-0.06934191,0.03855827,0.1201117,0.012023918,0.018013764,-0.009163978,-0.08580291,-0.04852857,0.010415257,-0.021901403,-0.0070036934,-0.120935075,0.0031510417,0.028975694,-0.043371145,0.005827526,0.028351381,0.056004304,-0.04193537,0.036470238,-0.02551061,0.05387047,0.041578542,-0.011544861,0.020083042,-0.019743534,-0.12281224,-0.03489822,-0.06302486,-0.004680995,0.0026240337,-0.049763817,0.03440226,0.006266021,0.021698607,0.011688389,-0.033272713,-0.005002312,0.010819187,-0.03326691,0.007720409,-0.015081273,0.02883106,-0.0051835533,0.03529145,0.0290507,0.059334554,0.043909762,-0.023158388,-5.2112195e-08,-0.02718794,0.03443475,0.030665524,0.073562786,0.04834766,-0.08066411,-0.027610883,-0.047683142,-0.027361598,-0.021350663,0.071528666,0.08550559,-0.044157986,0.00835718,0.036156062,0.03694163,-0.062062223,-0.06127477,-0.01934735,-0.031283908,0.08200542,-0.01270972,0.034525167,-0.040805645,0.035700735,-0.0045999265,0.0046523884,0.042981688,-0.012991904,0.121362366,0.018091332,-0.0011933425,-0.08453604,-0.023377812,0.0079048,-0.0118498225,-0.0031084812,-0.035165552,0.029298317,0.020293664,0.06833946,0.073024675,0.025958005,0.055193905,-0.0073304893,-0.08285115,0.012639976,-0.016050108,0.017499002,-0.070264466,-0.01374525,0.084155224,0.08962586,0.08847589,0.023961196,-0.028216874,-0.016020639,0.051627345,0.021587811,-0.017810969,0.09389319,-0.07359883,0.04477541,-0.040727414]

list_of_lists = [list_99249, list_7537, list_6028, list_141539, list_259669]
print(list_of_lists)

index_99249 = [99249]
index_7537 = [7537]
index_6028 = [6028]
index_141539 = [141539]
index_259669 = [259669]
indexes = index_99249 + index_7537 + index_6028 + index_141539 + index_259669

# Step 4: Calculate the cosine similarity matrix
similarity_matrix_list = cosine_similarity(list_of_lists, event_vectors)

# Step 5: Create df
similarity_df_list = pd.DataFrame(similarity_matrix_list,
                             index=indexes,
                             columns=event_embeddings['EventCode'])

similarity_df_list

similarity_df_list.T
