==> Jupyter Notebooks/6_EventsPerDate.ipynb <==
import pandas as pd
import gdown

import matplotlib.pyplot as plt

url_events_gdelt = 'https://drive.google.com/file/d/1hFhAQy2lucdIYBNSTL2fLR5syeT3utjw/view?usp=sharing'
path_gdelts_id = url_events_gdelt.split('/')[-2]
path_gdelt = f"https://drive.google.com/uc?id={path_gdelts_id}"
gdown.download(path_gdelt, 'gdelt_event_counts_1990_2024.csv', quiet=False)

# songs_embedded_df = pd.read_csv(path_songs)
gdelt_df = pd.read_csv('gdelt_event_counts_1990_2024.csv')

gdelt_df

gdelt_df.info()

gdelt_df['date'] = pd.to_datetime(gdelt_df['Year'].astype(str) + '-' + gdelt_df['Month'].astype(str) + '-01')

gdelt_df['date']

event_codes = gdelt_df['EventCode'].unique()

event_codes
