==> Jupyter Notebooks/7_Events_Artists_Genres.ipynb <==
import pandas as pd
import gdown
import matplotlib.pyplot as plt
import io

url_filtered_songs = 'https://drive.google.com/file/d/1zhzXgNwpfD-lPzZ4p8gJjyQCN4CGT6Na/view?usp=sharing'
path_filtered_songs_id = url_filtered_songs.split('/')[-2]
path_filtered_songs = f"https://drive.google.com/uc?id={path_filtered_songs_id}"
gdown.download(path_filtered_songs, 'songs_with_clustered_events_df.csv', quiet=False)

songs_filtered_df = pd.read_csv('songs_with_clustered_events_df.csv')

url_songs = 'https://drive.google.com/file/d/1pp6AwVMt2ROIMX96YPqkdmyDmMxA2G6w/view?usp=sharing'
path_songs_id = url_songs.split('/')[-2]
path_songs = f"https://drive.google.com/uc?id={path_songs_id}"
gdown.download(path_songs, 'songs_clean_text.csv', quiet=False)

songs_df = pd.read_csv('songs_clean_text.csv')

filtered_df = pd.merge(songs_df, songs_filtered_df, on='song_id', how='inner')
filtered_df

filtered_df.info()

filtered_df['Time_signature'].unique()

filtered_df['Instrumentalness'].unique()

columns_to_drop = [
    'song',
    'text',
    'Length',
    'Release_Date',
    'Liveness',
    'Acousticness',
    'Good_for_Party',
    'Good_for_Work_Study',
    'Good_for_Relaxation_Meditation',
    'Good_for_Exercise',
    'Good_for_Running',
    'Good_for_Yoga_Stretching',
    'Good_for_Driving',
    'Good_for_Social_Gatherings',
    'Good_for_Morning_Routine',
    'Similar_Artist_1',
    'Similar_Song_1',
    'Similarity_Score_1',
    'Similar_Artist_2',
    'Similar_Song_2',
    'Similarity_Score_2',
    'Similar_Artist_3',
    'Similar_Song_3',
    'Similarity_Score_3',
    # 'year',
    'month',
    'day',
    'decade',
    'language',
    'clean_text'
    ]

filtered_df = filtered_df.drop(columns_to_drop, axis=1)
filtered_df.info()

artist_frequency = filtered_df['Artists'].value_counts()
artist_frequency.head(10)

#top_artist = artist_frequency.idxmax()
#top_artist.head(5)

emotion_frequency = filtered_df['emotion'].value_counts()
emotion_frequency.head(10)

filtered_df['Genre'].unique()

filtered_df['Genre'].value_counts()

genre_frequency = filtered_df['Genre'].value_counts()
genre_frequency.head(10)

filtered_df['Genre_list'] = filtered_df['Genre'].str.split(',')

filtered_df['Genre_list']

def categorize_genres(genres_list):
    genres_str = ','.join(genres_list).lower()
    if any(g in genres_str for g in ['rock', 'emo', 'alternative rock']):
        return 'rock'
    elif any(g in genres_str for g in ['hardcore', 'punk', 'grunge']):
        return 'punk'
    elif any(g in genres_str for g in ['metal', 'metalcore', 'deathcore', 'nu metal']):
        return 'metal'
    elif any(g in genres_str for g in ['pop', 'indie pop', 'synthpop']):
        return 'pop'
    elif any(g in genres_str for g in ['hip hop', 'hip-hop', 'rap', 'grime', 'trap']):
        return 'hip hop'
    elif any(g in genres_str for g in ['country', 'folk', 'bluegrass']):
        return 'country'
    else:
        # return 'other'
        return genres_str

filtered_df['Genre_general'] = filtered_df['Genre_list'].apply(categorize_genres)

filtered_df['Genre_general'].value_counts().head(10)

def categorizar_generos(genres_str):
    if any(g in genres_str.lower() for g in ['rock', 'metal', 'hardcore', 'punk', 'grunge', 'emo']):
        return 'rock'
    elif any(g in genres_str.lower() for g in ['pop', 'indie pop', 'synthpop']):
        return 'pop'
    elif any(g in genres_str.lower() for g in ['hip hop', 'hip-hop', 'rap', 'grime']):
        return 'hip hop'
    elif any(g in genres_str.lower() for g in ['country', 'folk', 'bluegrass']):
        return 'country'
    # Añade más categorías según lo que necesites
    else:
        return 'otros'

album_frequency = filtered_df.value_counts(subset=['Album', 'Artists'])
album_frequency.head(10)
# filtered_album_frequency = album_frequency[album_frequency.index.get_level_values('year') == 2024]
# filtered_album_frequency.head(30)
